worker/requirements.txt
# Worker MQTT - Sistema de Seguridad IoT
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
